# upravljanjevremenom/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Time Manager API",
        "description": "REST API sistem za upravljanje vremenom. \nOmogućava korisnicima upravljanje zadacima, beleškama i obaveštenjima, \npraćenje statusa zadataka, kao i pregled statistike rada korisnika. \nSistem podržava uloge (admin, regular i premium korisnik).",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registracija korisnika (regular, premium ili admin)",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "tip"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jelena Vasiljević"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jelena@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "tip": {
                                        "description": "Tip korisnika u sistemu",
                                        "type": "string",
                                        "enum": [
                                            "regular",
                                            "premium",
                                            "admin"
                                        ],
                                        "example": "regular"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Uspešna registracija"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Prijava korisnika (Sanctum token)",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jelena@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspešna prijava"
                    },
                    "401": {
                        "description": "Neispravni kredencijali"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Odjava korisnika (briše trenutni token)",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Uspešna odjava"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Podaci o trenutno prijavljenom korisniku",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "Podaci o korisniku"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/beleske": {
            "get": {
                "tags": [
                    "Beleške"
                ],
                "summary": "Lista svih beleški",
                "operationId": "e9f6d51ea308d8e5670a7af7d6c75198",
                "responses": {
                    "200": {
                        "description": "Uspešno vraćena lista beleški"
                    }
                }
            },
            "post": {
                "tags": [
                    "Beleške"
                ],
                "summary": "Kreiranje nove beleške",
                "operationId": "ce504e2b69bed92f0e795494f5c38a80",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naslov",
                                    "sadrzaj",
                                    "korisnik_id"
                                ],
                                "properties": {
                                    "naslov": {
                                        "type": "string",
                                        "example": "Plan za danas"
                                    },
                                    "sadrzaj": {
                                        "type": "string",
                                        "example": "Završiti projekat i učiti za ispit."
                                    },
                                    "korisnik_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Beleška uspešno kreirana"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/beleske/{id}": {
            "get": {
                "tags": [
                    "Beleške"
                ],
                "summary": "Prikaz jedne beleške",
                "operationId": "945f96ecd9761404ea6bb1c06b1c1bfb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Beleška pronađena"
                    },
                    "404": {
                        "description": "Beleška nije pronađena"
                    }
                }
            },
            "put": {
                "tags": [
                    "Beleške"
                ],
                "summary": "Izmena beleške",
                "operationId": "ad793ece2d19d17ba0afe83f9c22977f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "naslov": {
                                        "type": "string",
                                        "example": "Novi naslov"
                                    },
                                    "sadrzaj": {
                                        "type": "string",
                                        "example": "Izmenjeni sadržaj"
                                    },
                                    "korisnik_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Beleška uspešno izmenjena"
                    },
                    "404": {
                        "description": "Beleška nije pronađena"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Beleške"
                ],
                "summary": "Brisanje beleške",
                "operationId": "020fff0d283b647b41637d2f232850a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Beleška uspešno obrisana"
                    },
                    "404": {
                        "description": "Beleška nije pronađena"
                    }
                }
            }
        },
...
    },
...
}